+++ src/main/java/zenpad/runners/exec/ClangExec.java
@@ -5,8 +5,8 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
 
-import zenpad.misc.DialogFactory;
 import zenpad.misc.LogUtils;
+import zenpad.misc.factory.DialogFactory;
 import zenpad.runners.TerminalLauncher;
 
 public class ClangExec implements CodeExec {
